Restructure project yet again
Yeah, I know I'm tinkering again, but I'm only wasting my own time here.
Not like anyone is paying me to write this code.
