+++ src/Algorithms.Core/DivideAndConquer/IInversionCounter.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Algorithms.Core.DivideAndConquer
+{
+    public interface IInversionCounter
+    {
+        long CountInversions(Span<int> integers);
+    }
+}
