+++ src/Algorithms.Core/DivideAndConquer/IKaratsuba.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Algorithms.Core.DivideAndConquer
+{
+    public interface IKaratsuba
+    {
+        Span<int> MultiplyXAndY(Span<int> x, Span<int> y);
+    }
+}
