+++ src/Algorithms/Algorithms.csproj
@@ -16,4 +16,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Algorithms.Core\Algorithms.Core.csproj" />
+  </ItemGroup>
+
 </Project>
