+++ src/Algorithms/Commands/CountInversionsCommand.cs
@@ -1,4 +1,5 @@
-﻿using Algorithms.Verbs;
+﻿using Algorithms.Core.DivideAndConquer;
+using Algorithms.Verbs;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -9,10 +10,12 @@ namespace Algorithms.Commands
     public class CountInversionsCommand : ICommand<CountInversionsOptions>
     {
         private readonly TextWriter _consoleOut;
+        private readonly IInversionCounter _inversionCounter;
 
-        public CountInversionsCommand(TextWriter consoleOut)
+        public CountInversionsCommand(TextWriter consoleOut, IInversionCounter inversionCounter)
         {
             _consoleOut = consoleOut;
+            _inversionCounter = inversionCounter;
         }
 
         public void Execute(CountInversionsOptions options)
@@ -21,7 +24,7 @@ namespace Algorithms.Commands
                 ? options.Integers
                 : ReadIntegerArrayFile(options.Path)).ToArray().AsSpan();
 
-            var inversions = CountInversions(integers);
+            var inversions = _inversionCounter.CountInversions(integers);
 
             _consoleOut.WriteLine(inversions);
         }
@@ -35,55 +38,6 @@ namespace Algorithms.Commands
             {
                 yield return int.Parse(line);
             }
-        }
-
-        private long CountInversions(Span<int> integers)
-        {
-            if (integers.Length <= 1)
-            {
-                return 0;
-            }
-            else
-            {
-                var left = integers.Slice(0, integers.Length / 2);
-                var right = integers.Slice(integers.Length / 2);
-                var leftInv = CountInversions(left);
-                var rightInv = CountInversions(right);
-                var splitInv = MergeAndCountSplitInv(left, right);
-
-                return leftInv + rightInv + splitInv;
-            }
-        }
-
-        private long MergeAndCountSplitInv(Span<int> left, Span<int> right)
-        {
-            var leftCtr = 0;
-            var rightCtr = 0;
-            var splitInv = 0;
-            var sorted = new int[left.Length + right.Length].AsSpan();
-
-            for (var ctr = 0; ctr < sorted.Length; ctr++)
-            {
-                if (leftCtr >= left.Length)
-                {
-                    sorted[ctr] = right[rightCtr++];
-                }
-                else if (rightCtr >= right.Length
-                    || left[leftCtr] < right[rightCtr])
-                {
-                    sorted[ctr] = left[leftCtr++];
-                }
-                else
-                {
-                    sorted[ctr] = right[rightCtr++];
-                    splitInv += left.Length - leftCtr;
-                }
-            }
-
-            sorted.Slice(0, left.Length).CopyTo(left);
-            sorted.Slice(left.Length).CopyTo(right);
-
-            return splitInv;
         }
     }
 }
