+++ test/Algorithms.Tests/Algorithms.Tests.csproj
@@ -29,6 +29,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Algorithms\Algorithms.csproj" />
+    <ProjectReference Include="..\..\src\Algorithms.Core\Algorithms.Core.csproj" />
   </ItemGroup>
 
 </Project>
